requirements.txt
alabaster==0.7.12
antiorm==1.2.1
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
colorama==0.4.1
cryptography==2.6.1
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.0
defusedxml==0.5.0
docutils==0.14
entrypoints==0.3
idna==2.5
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.17
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
keyring==18.0.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
newsapi-python==0.2.3
numpy==1.16.3
numpydoc==0.8.0
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psutil==5.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2019.1
pywin32==223
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
requests==2.17.1
rope==0.12.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.1
spyder==3.3.4
spyder-kernels==0.4.4
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.1
